+++ i18n/en.lua
@@ -341,6 +341,7 @@ local strings = {
     SI_LUIE_IGNORE_ERROR_FRIEND =                        "You cannot add a player you are ignoring as a friend.",
     SI_LUIE_IGNORE_ERROR_WHISPER =                       "You cannot whisper a player you are ignoring.",
     SI_LUIE_IGNORE_ERROR_GUILD =                         "You cannot extend a guild invitation to a player you are ignoring.",
+    SI_LUIE_IGNORE_ERROR_TRIBUTE =                       "You cannot play Tales of Tribute with a player you are ignoring.",
     SI_LUIE_NOTIFICATION_GROUP_INVITE =                  "Group invitation",
     SI_LUIE_NOTIFICATION_SHARE_QUEST_INVITE =            "Shared quest",
     SI_LUIE_NOTIFICATION_FRIEND_INVITE =                 "Friend invitation",